cleanup messaging logic
