+++ extension/src/background/navigation.ts
