+++ extension/src/frontend/view/ElementHighlight.tsx
