+++ extension/src/frontend/view/ElementInspector.hooks.ts
