+++ extension/src/frontend/view/ElementInspector.tsx
