+++ extension/src/frontend/view/InBrowserControls.tsx
