+++ extension/src/frontend/view/RemoteHighlights.hooks.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 
-import { client } from '../client'
+import { client } from '../routing'
 
 import { Bounds } from './types'
