+++ extension/src/frontend/view/RemoteHighlights.tsx
