+++ extension/src/frontend/view/hooks/useGlobalClass.ts
