+++ extension/src/frontend/view/index.tsx
