+++ extension/src/frontend/view/store.ts
