+++ extension/src/frontend/view/types.ts
