+++ extension/src/frontend/view/utils.ts
